ctlao/batch-order-generation: moved order generation to batch task service.
