+++ agent/src/main/java/com/cmclinnovations/agent/service/application/LifecycleTaskBatchService.java
@@ -1,30 +1,41 @@
 package com.cmclinnovations.agent.service.application;
 
+import java.util.ArrayDeque;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Queue;
 import java.util.Set;
+import java.util.stream.Collectors;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.cmclinnovations.agent.component.LocalisationTranslator;
 import com.cmclinnovations.agent.component.ResponseEntityBuilder;
+import com.cmclinnovations.agent.model.SparqlBinding;
+import com.cmclinnovations.agent.model.SparqlResponseField;
 import com.cmclinnovations.agent.model.response.StandardApiResponse;
 import com.cmclinnovations.agent.model.type.LifecycleEventType;
 import com.cmclinnovations.agent.model.type.LifecycleTaskOperationType;
 import com.cmclinnovations.agent.model.type.TrackActionType;
 import com.cmclinnovations.agent.service.AddService;
 import com.cmclinnovations.agent.service.DeleteService;
+import com.cmclinnovations.agent.service.GetService;
 import com.cmclinnovations.agent.service.core.ChangelogService;
 import com.cmclinnovations.agent.service.core.DateTimeService;
 import com.cmclinnovations.agent.service.core.FileService;
+import com.cmclinnovations.agent.template.LifecycleQueryFactory;
 import com.cmclinnovations.agent.utils.LifecycleResource;
 import com.cmclinnovations.agent.utils.LocalisationResource;
 import com.cmclinnovations.agent.utils.QueryResource;
+import com.cmclinnovations.agent.utils.StringResource;
+import com.cmclinnovations.agent.utils.TypeCastUtils;
 
 @Service
 public class LifecycleTaskBatchService {
@@ -32,20 +43,29 @@ public class LifecycleTaskBatchService {
   private final ChangelogService changelogService;
   private final DateTimeService dateTimeService;
   private final DeleteService deleteService;
+  private final GetService getService;
   private final LifecycleQueryService lifecycleQueryService;
   private final LifecycleTaskService lifecycleTaskService;
   private final ResponseEntityBuilder responseEntityBuilder;
 
+  private final LifecycleQueryFactory lifecycleQueryFactory;
+
+  private static final int NUM_DAY_ORDER_GEN = 30;
+  private static final Logger LOGGER = LogManager.getLogger(LifecycleTaskBatchService.class);
+
   public LifecycleTaskBatchService(AddService addService, ChangelogService changelogService,
-      DateTimeService dateTimeService, DeleteService deleteService, LifecycleQueryService lifecycleQueryService,
-      LifecycleTaskService lifecycleTaskService, ResponseEntityBuilder responseEntityBuilder) {
+      DateTimeService dateTimeService, DeleteService deleteService, GetService getService,
+      LifecycleQueryService lifecycleQueryService, LifecycleTaskService lifecycleTaskService,
+      ResponseEntityBuilder responseEntityBuilder) {
     this.addService = addService;
     this.changelogService = changelogService;
     this.dateTimeService = dateTimeService;
     this.deleteService = deleteService;
+    this.getService = getService;
     this.lifecycleQueryService = lifecycleQueryService;
     this.lifecycleTaskService = lifecycleTaskService;
     this.responseEntityBuilder = responseEntityBuilder;
+    this.lifecycleQueryFactory = new LifecycleQueryFactory();
   }
 
   /**
@@ -218,5 +238,100 @@ public class LifecycleTaskBatchService {
         item.put(LifecycleResource.EVENT_STATUS_KEY, config.getEventStatus());
       }
     }
+  }
+
+  /**
+   * Check for active contract and generate orders up to the limit date.
+   */
+  public void genOrderActiveContracts() {
+    String todayString = this.dateTimeService.getCurrentDate();
+    String taskGenerationCutoffDate = this.dateTimeService.getFutureDate(todayString, NUM_DAY_ORDER_GEN);
+    LOGGER.info("Retrieving all active contracts that need orders to be generated...");
+    String query = this.lifecycleQueryFactory.getLatestOrderQuery(taskGenerationCutoffDate);
+    Queue<SparqlBinding> results = this.getService.getInstances(query);
+    while (!results.isEmpty()) {
+      SparqlBinding resultRow = results.poll();
+      String currentContract = resultRow.getFieldValue(QueryResource.ID_KEY);
+      // Latest task date for the contract
+      String latestTaskDate = resultRow.getFieldValue(QueryResource.LATEST_DATE_VAR.getVarName());
+      String nextTaskStartDate = this.dateTimeService.getFutureDate(latestTaskDate, 1);
+      LOGGER.info("Generating orders for contract {}, starting from {}", currentContract, nextTaskStartDate);
+      this.genOrderReceivedOccurrences(currentContract, nextTaskStartDate);
+    }
+  }
+
+  /**
+   * Generate occurrences for the order received event of a specified contract.
+   * 
+   * @param contract          Target contract.
+   * @param nextTaskStartDate Optional parameter that indicates the next task
+   *                          start date. If provided, this will overwrite the
+   *                          contract start date.
+   * @return boolean indicating if the occurrences have been generated
+   *         successfully.
+   */
+  public boolean genOrderReceivedOccurrences(String contract, String nextTaskStartDate) {
+    LOGGER.info("Generating all orders for the active contract {}...", contract);
+    // Retrieve schedule information for the specific contract
+    SparqlBinding bindings = this.lifecycleQueryService.querySchedule(contract);
+    // Extract specific schedule info
+    String startDate = nextTaskStartDate != null ? nextTaskStartDate
+        : bindings.getFieldValue(QueryResource.SCHEDULE_START_DATE_VAR.getVarName());
+    // For non-perpetual schedules, get earliest date cutoff or contract end date
+    String endDate = null;
+    if (bindings.containsField(QueryResource.SCHEDULE_END_DATE_VAR.getVarName())) {
+      String endDateVal = bindings.getFieldValue(QueryResource.SCHEDULE_END_DATE_VAR.getVarName());
+      endDate = this.dateTimeService.getEarliestDateOrContractEnd(endDateVal, NUM_DAY_ORDER_GEN);
+    }
+    String recurrence = bindings.getFieldValue(LifecycleResource.SCHEDULE_RECURRENCE_PLACEHOLDER_KEY);
+    Queue<String> occurrences = new ArrayDeque<>();
+    // Handle as fixed date schedule first
+    if (bindings.containsField(QueryResource.FIXED_DATE_DATE_KEY)) {
+      List<Map<String, SparqlResponseField>> entryDates = bindings.getList(QueryResource.FIXED_DATE_DATE_KEY);
+      List<String> entryDateStrings = entryDates.stream()
+          .map(entryDate -> entryDate.get(QueryResource.FIXED_DATE_DATE_KEY).value())
+          .collect(Collectors.toList());
+      occurrences = this.dateTimeService.getOccurrenceDates(entryDateStrings, endDate);
+    } else {
+      // Extract date of occurrences based on the schedule information
+      // For perpetual and single time schedules, simply add the start date
+      if (recurrence == null || recurrence.equals(LifecycleResource.RECURRENCE_DAILY_TASK)) {
+        occurrences.offer(this.dateTimeService.getDateTimeFromDate(startDate));
+      } else if (recurrence.equals(LifecycleResource.RECURRENCE_ALT_DAY_TASK)) {
+        // Alternate day recurrence should have dual interval
+        occurrences = this.dateTimeService.getOccurrenceDates(startDate, endDate, 2);
+      } else {
+        // Note that this may run for other intervals like P3D but
+        // an error will be thrown in the following method unless the recurrence is in
+        // intervals of 7
+        int weeklyInterval = this.dateTimeService.getWeeklyInterval(recurrence);
+        occurrences = this.dateTimeService.getOccurrenceDates(startDate, endDate, bindings, weeklyInterval);
+      }
+    }
+    // Add parameter template
+    Map<String, Object> params = new HashMap<>();
+    params.put(LifecycleResource.CONTRACT_KEY, contract);
+    params.put(LifecycleResource.REMARKS_KEY, LifecycleResource.ORDER_INITIALISE_MESSAGE);
+    this.lifecycleQueryService.addOccurrenceParams(params, LifecycleEventType.SERVICE_ORDER_RECEIVED);
+    String orderPrefix = StringResource.getPrefix(params.get(LifecycleResource.STAGE_KEY).toString());
+    // Instantiate each occurrence
+    boolean hasError = false;
+    while (!occurrences.isEmpty()) {
+      // Retrieve and update the date of occurrence
+      String occurrenceDate = occurrences.poll();
+      // set new id each time
+      params.remove(QueryResource.ID_KEY);
+      LifecycleResource.genIdAndInstanceParameters(orderPrefix, LifecycleEventType.SERVICE_ORDER_RECEIVED, params);
+      params.put(LifecycleResource.DATE_TIME_KEY, occurrenceDate);
+      try {
+        this.addService.instantiate(LifecycleResource.OCCURRENCE_INSTANT_RESOURCE, params, TrackActionType.CREATION);
+        // Error logs for any specified occurrence
+      } catch (IllegalStateException _) {
+        LOGGER.error("Error encountered while creating order for {} on {}! Read error logs for more details",
+            contract, occurrenceDate);
+        hasError = true;
+      }
+    }
+    return hasError;
   }
 }
