+++ agent/src/main/java/com/cmclinnovations/agent/service/application/LifecycleTaskService.java
@@ -884,101 +884,6 @@ public class LifecycleTaskService {
   }
 
   /**
-   * Check for active contract and generate orders up to the limit date.
-   */
-  public void genOrderActiveContracts() {
-    String todayString = this.dateTimeService.getCurrentDate();
-    String taskGenerationCutoffDate = this.dateTimeService.getFutureDate(todayString, NUM_DAY_ORDER_GEN);
-    LOGGER.info("Retrieving all active contracts that need orders to be generated...");
-    String query = this.lifecycleQueryFactory.getLatestOrderQuery(taskGenerationCutoffDate);
-    Queue<SparqlBinding> results = this.getService.getInstances(query);
-    while (!results.isEmpty()) {
-      SparqlBinding resultRow = results.poll();
-      String currentContract = resultRow.getFieldValue(QueryResource.ID_KEY);
-      // Latest task date for the contract
-      String latestTaskDate = resultRow.getFieldValue(QueryResource.LATEST_DATE_VAR.getVarName());
-      String nextTaskStartDate = this.dateTimeService.getFutureDate(latestTaskDate, 1);
-      LOGGER.info("Generating orders for contract {}, starting from {}", currentContract, nextTaskStartDate);
-      this.genOrderReceivedOccurrences(currentContract, nextTaskStartDate);
-    }
-  }
-
-  /**
-   * Generate occurrences for the order received event of a specified contract.
-   * 
-   * @param contract          Target contract.
-   * @param nextTaskStartDate Optional parameter that indicates the next task
-   *                          start date. If provided, this will overwrite the
-   *                          contract start date.
-   * @return boolean indicating if the occurrences have been generated
-   *         successfully.
-   */
-  public boolean genOrderReceivedOccurrences(String contract, String nextTaskStartDate) {
-    LOGGER.info("Generating all orders for the active contract {}...", contract);
-    // Retrieve schedule information for the specific contract
-    SparqlBinding bindings = this.lifecycleQueryService.querySchedule(contract);
-    // Extract specific schedule info
-    String startDate = nextTaskStartDate != null ? nextTaskStartDate
-        : bindings.getFieldValue(QueryResource.SCHEDULE_START_DATE_VAR.getVarName());
-    // For non-perpetual schedules, get earliest date cutoff or contract end date
-    String endDate = null;
-    if (bindings.containsField(QueryResource.SCHEDULE_END_DATE_VAR.getVarName())) {
-      String endDateVal = bindings.getFieldValue(QueryResource.SCHEDULE_END_DATE_VAR.getVarName());
-      endDate = this.dateTimeService.getEarliestDateOrContractEnd(endDateVal, NUM_DAY_ORDER_GEN);
-    }
-    String recurrence = bindings.getFieldValue(LifecycleResource.SCHEDULE_RECURRENCE_PLACEHOLDER_KEY);
-    Queue<String> occurrences = new ArrayDeque<>();
-    // Handle as fixed date schedule first
-    if (bindings.containsField(QueryResource.FIXED_DATE_DATE_KEY)) {
-      List<Map<String, SparqlResponseField>> entryDates = bindings.getList(QueryResource.FIXED_DATE_DATE_KEY);
-      List<String> entryDateStrings = entryDates.stream()
-          .map(entryDate -> entryDate.get(QueryResource.FIXED_DATE_DATE_KEY).value())
-          .collect(Collectors.toList());
-      occurrences = this.dateTimeService.getOccurrenceDates(entryDateStrings, endDate);
-    } else {
-      // Extract date of occurrences based on the schedule information
-      // For perpetual and single time schedules, simply add the start date
-      if (recurrence == null || recurrence.equals(LifecycleResource.RECURRENCE_DAILY_TASK)) {
-        occurrences.offer(this.dateTimeService.getDateTimeFromDate(startDate));
-      } else if (recurrence.equals(LifecycleResource.RECURRENCE_ALT_DAY_TASK)) {
-        // Alternate day recurrence should have dual interval
-        occurrences = this.dateTimeService.getOccurrenceDates(startDate, endDate, 2);
-      } else {
-        // Note that this may run for other intervals like P3D but
-        // an error will be thrown in the following method unless the recurrence is in
-        // intervals of 7
-        int weeklyInterval = this.dateTimeService.getWeeklyInterval(recurrence);
-        occurrences = this.dateTimeService.getOccurrenceDates(startDate, endDate, bindings, weeklyInterval);
-      }
-    }
-    // Add parameter template
-    Map<String, Object> params = new HashMap<>();
-    params.put(LifecycleResource.CONTRACT_KEY, contract);
-    params.put(LifecycleResource.REMARKS_KEY, LifecycleResource.ORDER_INITIALISE_MESSAGE);
-    this.lifecycleQueryService.addOccurrenceParams(params, LifecycleEventType.SERVICE_ORDER_RECEIVED);
-    String orderPrefix = StringResource.getPrefix(params.get(LifecycleResource.STAGE_KEY).toString());
-    // Instantiate each occurrence
-    boolean hasError = false;
-    while (!occurrences.isEmpty()) {
-      // Retrieve and update the date of occurrence
-      String occurrenceDate = occurrences.poll();
-      // set new id each time
-      params.remove(QueryResource.ID_KEY);
-      LifecycleResource.genIdAndInstanceParameters(orderPrefix, LifecycleEventType.SERVICE_ORDER_RECEIVED, params);
-      params.put(LifecycleResource.DATE_TIME_KEY, occurrenceDate);
-      try {
-        this.addService.instantiate(LifecycleResource.OCCURRENCE_INSTANT_RESOURCE, params, TrackActionType.CREATION);
-        // Error logs for any specified occurrence
-      } catch (IllegalStateException _) {
-        LOGGER.error("Error encountered while creating order for {} on {}! Read error logs for more details",
-            contract, occurrenceDate);
-        hasError = true;
-      }
-    }
-    return hasError;
-  }
-
-  /**
    * Continues the task today by generating a new occurrence for the order
    * received and dispatch event of a specified contract.
    *
